added check for double

+++ StoredData.cs
@@ -5,7 +5,7 @@ public class StoredData
     public ReadAPIData APIreader = new ReadAPIData();
     public string? dataRead {  get; set; }
     public Root? myDeserializedData { get; set; }
-    public Dictionary<string, (int? Population, int? Diameter, int? SurfaceWater)> Planets { get; set; } = new();
+    public Dictionary<string, (double? Population, int? Diameter, int? SurfaceWater)> Planets { get; set; } = new();
     public async Task getDataAndStoreIndataRead()
     {
         dataRead = await APIreader.readData(APIreader.baseURL, APIreader.RestofURL);
@@ -20,7 +20,7 @@ public class StoredData
         if(myDeserializedData == null) return;
         foreach (var planet in myDeserializedData.results)
         {
-            Planets[planet.Name] = (Population: ParseInput.ParseInt(planet.Population), 
+            Planets[planet.Name] = (Population: ParseInput.ParseDouble(planet.Population), 
                 Diameter: ParseInput.ParseInt(planet.Diameter),
                 SurfaceWater: ParseInput.ParseInt(planet.SurfaceWater));
         }
